pushing in 28/6

+++ database/migrations/2022_04_08_140612_create_usermodel2s_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
                 ->references('id')
                 ->on('users')
                 ->onDelete('cascade');
-            $table->integer('evaluation')->nullable();
+            $table->float('evaluation')->nullable();
             $table->integer('salary')->nullable();
             $table->rememberToken();
             $table->timestamps();
